Improve performance when clicking on an item
It would appear that this condition can be called a lot of times and then create performances issues when clicking on an item of a list (basically an email). The slow performance seems to come from the "$(this.container).is(':visible')" test which can take several milliseconds, multiplied by a lot of iterations in the loop (it probably depends on other plugins as this event is triggers every time a plugin call get_selection) it may take a few seconds to open a message.

But, most of the time "this.menu_selection.length" seem to be equal to 0, so putting the the lightest condition first brings better performance and avoid to use the poor performance "is visible" test.